move implementation into extractComments()
